Use new handle_uncaught_exceptions in start_qapplication
New exception hook from gui system tests is better for catching
exceptions in slots and reporting rather than exiting. Use it for all
tests that use the start_qapplication decorator.

This also allows GuiSystemBase to be subclassed multiple times.
