+++ mantidimaging/gui/test/gui_system_base.py
@@ -3,7 +3,6 @@
 
 import os
 from pathlib import Path
-import sys
 import unittest
 from unittest import mock
 
@@ -15,6 +14,7 @@ import pytest
 from mantidimaging.core.utility.version_check import versions
 from mantidimaging.gui.windows.main import MainWindowView
 from mantidimaging.gui.windows.load_dialog.presenter import Notification
+from mantidimaging.test_helpers.start_qapplication import start_qapplication
 
 versions._use_test_values()
 
@@ -26,37 +26,14 @@ SHOW_DELAY = 10  # Can be increased to watch tests
 SHORT_DELAY = 100
 LOAD_DELAY = 5000
 
-uncaught_exception = None
-current_excepthook = sys.excepthook
-
-
-def handle_uncaught_exceptions(exc_type, exc_value, exc_traceback):
-    """
-    Qt slots swallows exceptions. We need to catch them, but not exit.
-    """
-    global uncaught_exception
-    # store first exception caught
-    if uncaught_exception is None:
-        uncaught_exception = f"{exc_type=}, {exc_value=}"
-
-    current_excepthook(exc_type, exc_value, exc_traceback)
-
-
-sys.excepthook = handle_uncaught_exceptions
-
 
 @pytest.mark.system
 @unittest.skipUnless(os.path.exists(LOAD_SAMPLE), LOAD_SAMPLE_MISSING_MESSAGE)
+@start_qapplication
 class GuiSystemBase(unittest.TestCase):
     app: QApplication
 
-    @classmethod
-    def setUpClass(cls):
-        cls.app = QApplication([])
-
     def setUp(self) -> None:
-        global uncaught_exception
-        uncaught_exception = None
         self.main_window = MainWindowView()
         self.main_window.show()
         QTest.qWait(SHORT_DELAY)
@@ -65,9 +42,6 @@ class GuiSystemBase(unittest.TestCase):
         QTimer.singleShot(SHORT_DELAY, lambda: self._click_messageBox("Yes"))
         self.main_window.close()
         QTest.qWait(SHORT_DELAY)
-
-        if uncaught_exception is not None:
-            pytest.fail(f"Uncaught exception {uncaught_exception}")
 
     @classmethod
     def _click_messageBox(cls, button_text: str):
